Made a quick environment for self play.
Need to test the learning method before making the plugin obviously. I think there should be a reward for blocking an opponent. Might rework the agent brain, but I'll keep this one until I can prove that different types of reward schemes won't help. But about 15000 epochs was not good enough to beat a human on a regular basis.